+++ testapp/tests/test_e2e_richtextarea.py
@@ -7,7 +7,7 @@ from django.forms import fields, forms
 from django.urls import path
 
 from formset.richtext import controls, dialogs
-from formset.richtext.widgets import RichTextarea
+from formset.widgets import RichTextarea
 from formset.views import FormView
 
 from .utils import ContextMixin, get_javascript_catalog